+++ R/dataset.R
@@ -5,8 +5,8 @@
 #' to demonstrate the functions in `stcvlite`.
 #'
 #' @family Dataset
-#' @format A data frame with 12,000 rows (100 locations * 120 times) and
-#'   three variables:
+#' @format A `data.table` with 12,000 rows (100 locations * 120 times) and
+#'   six variables, plus a `"crs"` attribute (`attr(spdat, "crs")`):
 #' \describe{
 #' \item{time}{Unique point identifier. Arbitrarily generated.}
 #' \item{lon}{Longitude}